+++ src/main/java/com/kosa/mapbegood/domain/common/entity/AuditEntity.java
@@ -9,9 +9,6 @@ import javax.persistence.MappedSuperclass;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import com.kosa.mapbegood.domain.favorite.entity.FavoriteEmbedded;
-
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
 @Getter
